requirements.txt
streamlit>=1.33
pandas>=2.0
numpy>=1.23
plotly>=5.15
python-dotenv>=1.0
nixtla>=0.5.1